requirements.txt
font_fredoka_one
Pillow
inky
numpy